gst/vaapi: move MPEG2 encoder test classes to lib
