+++ test/gst-vaapi/encode/mpeg2.py
@@ -5,30 +5,11 @@
 ###
 
 from ....lib import *
-from ....lib.codecs import Codec
 from ....lib.gstreamer.vaapi.util import *
-from ....lib.gstreamer.vaapi.encoder import EncoderTest
+from ....lib.gstreamer.vaapi.encoder import MPEG2EncoderTest
 
 spec      = load_test_spec("mpeg2", "encode")
 spec_r2r  = load_test_spec("mpeg2", "encode", "r2r")
-
-@slash.requires(*have_gst_element("vaapimpeg2enc"))
-@slash.requires(*have_gst_element("vaapimpeg2dec"))
-@slash.requires(*platform.have_caps("encode", "mpeg2"))
-class MPEG2EncoderTest(EncoderTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      caps          = platform.get_caps("encode", "mpeg2"),
-      codec         = Codec.MPEG2,
-      gstencoder    = "vaapimpeg2enc",
-      gstdecoder    = "vaapimpeg2dec",
-      gstmediatype  = "video/mpeg,mpegversion=2",
-      gstparser     = "mpegvideoparse",
-    )
-
-  def get_file_ext(self):
-    return "m2v"
 
 class cqp(MPEG2EncoderTest):
   def init(self, tspec, case, gop, bframes, qp, quality):
